[bugFix] - node with a longer term than the network but not leader
A node participating in the network with a term equal to that of the network does not receive the heartbeat and
enters the election by increasing its term.

However, the other nodes received all the heartbeats and therefore have the variable votedfor filled in,
as a consequence they did not choose the new node even if it had a term higher than that of the current leader.

Now the nodes accept a new leader whenever it has a higher term than the network, regardless of whether the votedfor
variable is filled or not.