[bugFix] - fixed strategy to suspend the application
Suspending the application through the operating system (kill-STOP) does not reflect a total communication failure as the data
is buffered during freezing and is available to the application after the resume.

In this new approach a complete freeze is simulated. The received data is discarded and the timer that triggers the timeout
is paused, but unlike the reset, the RAM memory of the node is kept intact, that is, a leader will still think he is a leader
after the resume.